requirements.txt
alabaster==0.7.9
djangoajax==2.3.6
argh==0.26.2
Babel==2.3.4
backports-abc==0.4
cairocffi==0.7.2
CairoSVG==1.0.22
cffi==1.7.0
cssselect==0.9.2
Django>=1.11
docutils==0.12
html5lib==0.999999999
imagesize==0.7.1
Jinja2==2.8
livereload==2.4.1
lxml==3.6.1
MarkupSafe==0.23
pathtools==0.1.2
Pillow==3.3.0
port-for==0.3.1
psycopg2==2.6.2
pycparser==2.14
Pygments==2.1.3
Pyphen==0.9.4
pytz==2016.6.1
PyYAML==3.11
reportlab==3.3.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.5
sphinx-autobuild==0.6.0
sphinx-rtd-theme==0.1.9
tinycss==0.3
tornado==4.4.1
watchdog==0.8.3
WeasyPrint==0.30
webencodings==0.5
simplejson==3.8.2
django-jsonfield==1.0.1
django-json-rpc==0.6.1
django-bootstrap-form
pyEQL==0.4
django-reservation==0.2.8
celery==4.0.1
redis==2.10.5
django-ajax-selects==1.5.2
django-cruds-adminlte>=0.0.9
pyexcel==0.5.6
pyexcel-ezodf==0.3.4
pyexcel-io==0.5.4
pyexcel-ods3==0.5.2
pyexcel-xls==0.5.4
django-excel==0.0.9
pyexcel-xlsx==0.5.4
django-chartjs==1.2
